+++ lightly_studio/tests/resolvers/annotations/test_annotations_base_get_all_annotations_details.py
@@ -86,7 +86,7 @@ def test_default_ordering_by_file_path_abs(
     """Test that annotations are ordered by sample file path."""
     annotation_label = create_annotation_label(
         session=db_session,
-        root_collection_id=collection.collection_id,
+        dataset_id=collection.collection_id,
         label_name="cat",
     )
     # Create samples in random order.