+++ lightly_studio/tests/resolvers/datasets_resolver/test_export.py
@@ -42,12 +42,12 @@ def test_collection_export(test_db: Session) -> TestcollectionExport:
     # create annotation_tag
     cat_label = create_annotation_label(
         session=test_db,
-        root_collection_id=collection_id,
+        dataset_id=collection_id,
         label_name="cat",
     )
     dog_label = create_annotation_label(
         session=test_db,
-        root_collection_id=collection_id,
+        dataset_id=collection_id,
         label_name="dog",
     )
 
@@ -170,7 +170,7 @@ def test_collection_export(test_db: Session) -> TestcollectionExport:
     )
     parrot_label = create_annotation_label(
         session=test_db,
-        root_collection_id=collection2.collection_id,
+        dataset_id=collection2.collection_id,
         label_name="parrot",
     )
     create_annotation(
@@ -719,7 +719,7 @@ def test_export__exclude_by_annotation_id__ensure_samples_without_annotations_ar
     )
     parrot_label = create_annotation_label(
         session=test_db,
-        root_collection_id=collection.collection_id,
+        dataset_id=collection.collection_id,
         label_name="parrot",
     )
     # create annotaitons only for sample 2