+++ lightly_studio/tests/resolvers/datasets_resolver/test_get_parent_dataset_id.py
@@ -21,7 +21,7 @@ def test_get_parent_collection_id__from_parent_collection(test_db: Session) -> N
     )
     car_label = create_annotation_label(
         session=test_db,
-        root_collection_id=collection.collection_id,
+        dataset_id=collection.collection_id,
         label_name="car",
     )
     annotation = create_annotation(