+++ lightly_studio/tests/resolvers/video/video_resolver/test_get_all_by_dataset_id.py
@@ -212,13 +212,13 @@ def test_get_all_by_collection_id__with_annotation_frames_label_filter(
 
     car_label = create_annotation_label(
         session=test_db,
-        root_collection_id=collection_id,
+        dataset_id=collection_id,
         label_name="car",
     )
 
     airplane_label = create_annotation_label(
         session=test_db,
-        root_collection_id=collection_id,
+        dataset_id=collection_id,
         label_name="airplane",
     )