+++ lightly_studio_view/src/lib/components/NavigationMenu/types.ts
@@ -6,4 +6,5 @@ export type NavigationMenuItem = {
     isSelected: boolean;
     icon?: Component;
     id: string;
+    children?: NavigationMenuItem[];
 };